Refactored Text Generation files
