+++ generative_ai/image/image_example01.py
