+++ generative_ai/image/image_example02.py
