+++ generative_ai/image/test_image_samples.py
@@ -1,0 +1,27 @@
+# Copyright 2024 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import image_example01
+import image_example02
+
+
+def test_gemini_guide_example() -> None:
+    text = image_example01.generate_text()
+    text = text.lower()
+    assert len(text) > 0
+
+
+def test_gemini_pro_basic_example() -> None:
+    text = image_example02.generate_text()
+    assert len(text) > 0
