+++ generative_ai/openai/chat_openai_example.py
