+++ generative_ai/text_generation/test_text_generation_examples.py
@@ -14,23 +14,29 @@
 import os
 
 import backoff
-
+import chat_code_example
+import chat_multiturn_example
+import chat_multiturn_stream_example
+import chat_openai_image_example
+import chat_openai_image_stream_example
+import chat_openai_stream_example
+import chat_simple_example
 import code_completion_example
 import codegen_example
-import credentials_refresher_usage_example
 import gemini_describe_http_image_example
 import gemini_describe_http_pdf_example
+import generation_config_example
 
 from google.api_core.exceptions import ResourceExhausted
 
-import image_example01
-import image_example02
-import multimodal_example01
-import multimodal_example02
 import multimodal_stream_example
-import pdf_example
-import pro_config_example
 import single_turn_multi_image_example
+
+
+@backoff.on_exception(backoff.expo, ResourceExhausted, max_time=10)
+def test_code_chat() -> None:
+    content = chat_code_example.write_a_function().text
+    assert len(content) > 0
 
 
 def test_gemini_describe_http_image_example() -> None:
@@ -49,21 +55,6 @@ def test_code_completion_comment() -> None:
     assert len(content) > 0
 
 
-def test_credentials_refresher() -> None:
-    response = credentials_refresher_usage_example.generate_text()
-    assert response
-
-
-def test_analyze_all_modalities() -> None:
-    text = multimodal_example01.analyze_all_modalities()
-    assert len(text) > 0
-
-
-def test_stream_multi_modality_basic() -> None:
-    responses = multimodal_example02.generate_content()
-    assert responses
-
-
 def test_stream_multi_modality_basic_example() -> None:
     responses = multimodal_stream_example.generate_content()
     assert responses
@@ -77,12 +68,6 @@ def test_code_generation_function() -> None:
     assert "return" in content
 
 
-def test_gemini_guide_example() -> None:
-    text = image_example01.generate_text()
-    text = text.lower()
-    assert len(text) > 0
-
-
 def test_gemini_multi_image_example() -> None:
     text = single_turn_multi_image_example.generate_text_multimodal()
     text = text.lower()
@@ -91,29 +76,16 @@ def test_gemini_multi_image_example() -> None:
     assert "landmark" in text
 
 
-# @pytest.mark.skip(
-#     "TODO: Exception Logs indicate safety filters are likely blocking model output b/339985493"
-# )
-def test_gemini_pdf_example() -> None:
-    text = pdf_example.analyze_pdf()
-    assert len(text) > 0
-
-
-def test_gemini_pro_basic_example() -> None:
-    text = image_example02.generate_text()
-    assert len(text) > 0
-
-
 def test_gemini_pro_config_example() -> None:
     import urllib.request
 
-    #  download the image
+    # Download the image
     fname = "scones.jpg"
     url = "https://storage.googleapis.com/generativeai-downloads/images/scones.jpg"
     urllib.request.urlretrieve(url, fname)
 
     if os.path.isfile(fname):
-        text = pro_config_example.generate_text()
+        text = generation_config_example.generate_text()
         text = text.lower()
         assert len(text) > 0
 
@@ -121,3 +93,36 @@ def test_gemini_pro_config_example() -> None:
         os.remove(fname)
     else:
         raise Exception("File(scones.jpg) not found!")
+
+
+@backoff.on_exception(backoff.expo, ResourceExhausted, max_time=10)
+def test_chat_example() -> None:
+    response = chat_simple_example.send_chat()
+    assert len(response) > 0
+
+
+def test_gemini_chat_example() -> None:
+    text = chat_multiturn_example.chat_text_example()
+    text = text.lower()
+    assert len(text) > 0
+    assert any([_ in text for _ in ("hi", "hello", "greeting")])
+
+    text = chat_multiturn_stream_example.chat_stream_example()
+    text = text.lower()
+    assert len(text) > 0
+    assert any([_ in text for _ in ("hi", "hello", "greeting")])
+
+
+def test_non_streaming_image() -> None:
+    response = chat_openai_image_example.generate_text()
+    assert response
+
+
+def test_streaming_image() -> None:
+    response = chat_openai_image_stream_example.generate_text()
+    assert response
+
+
+def test_streaming_text() -> None:
+    response = chat_openai_stream_example.generate_text()
+    assert response
