+++ generative_ai/video/multimodal_example01.py
