+++ generative_ai/video/multimodal_example02.py
